// app/dashboard/employees/types/employeesType.ts
export type User = {
  id: number;
  firstName: string;
  lastName: string;
  maidenName?: string;
  age: number;
  gender: "male" | "female";
  email: string;
  phone: string;
  username: string;
  password: string;
  birthDate: string;
  image: string;
  bloodGroup: string;
  height: number;
  weight: number;
  eyeColor: string;
  hair: {
    color: string;
    type: string;
  };
  ip: string;
  address: {
    address: string;
    city: string;
    state: string;
    stateCode: string;
    postalCode: string;
    coordinates: {
      lat: number;
      lng: number;
    };
    country: string;
  };
  macAddress: string;
  university: string;
  bank: {
    cardExpire: string;
    cardNumber: string;
    cardType: string;
    currency: string;
    iban: string;
  };
  company: {
    department: string;
    name: string;
    title: string;
    address: {
      address: string;
      city: string;
      state: string;
      stateCode: string;
      postalCode: string;
      coordinates: {
        lat: number;
        lng: number;
      };
      country: string;
    };
  };
  ein: string;
  ssn: string;
  userAgent: string;
  crypto: {
    coin: string;
    wallet: string;
    network: string;
  };
  role: "admin" | "moderator" | "user";
};
export type ApiResponse<T> = {
  users: T[];
  total: number;
  skip: number;
  limit: number;
};

export interface SingleUserType {
  id: string;
  fullName: string;
  email: string;
  phoneNumber: number;
  joinDate: string;
  avatar: string;
  permissions: string[];
  password: string;
  status: "ACTIVE" | "INACTIVE" | "SUSPENDED";
  role: "ADMIN" | "MODERATOR" | "CLIENT";
  createdAt: string;
  address: Address[];
  Profile: {
    profilePhoto?: string;
    gender?: string;
    profileId: string;
    dateOfBirth: string | null;
    department: string | null;
    experience: string | null;
    licenseNo: string | null;
    age: number | null;
    weight: number | null;
    height: number | null;
    bloodGroup: string | null;
    licenseFile: string[]; // Assuming URLs or file names
    preferredLocations: string[];
    createdAt: string;
    updatedAt: string;
    userId: string;
  };
  profile?: {
    profilePhoto?: string;
    gender?: string;
    profileId: string;
    dateOfBirth: string | null;
    department: string | null;
    experience: string | null;
    licenseNo: string | null;
    age: number | null;
    weight: number | null;
    height: number | null;
    bloodGroup: string | null;
    licenseFile: string[];
    preferredLocations: string[];
    createdAt: string;
    updatedAt: string;
    userId: string;
  };
}